requirements.txt
azure-identity==1.16.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.20.0
pyodbc==5.1.0
python-dotenv==1.0.1
python-docx==1.1.2
pypdf==4.3.1
openai==1.44.0
tiktoken==0.7.0
